pyproject.toml
[tool.poetry]
name = "django-business-management-practice"
version = "0.1.0"
description = ""
authors = ["exkuretrol <33695301+exkuretrol@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
crispy-bootstrap5 = "^2024.2"
django-crispy-forms = "^2.1"
rich = "^13.7.1"
django-tables2 = "^2.7.0"
django-quill-editor = "^0.1.40"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
djlint = "^1.34.1"
django-webpack-loader = "^3.1.0"
django-filter = "^24.2"
django-debug-toolbar = "^4.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
